Move DTOs to shared package per code review
- Move CreateTaskDto, QueryTaskDto, UpdateTaskDto to packages/shared/src/dto
- Export DTOs from shared package index
- Update imports in TasksService and TasksController to use @shared/types
- Update test imports to use @shared/types
- Add class-validator and class-transformer dependencies to shared package
- Configure shared package tsconfig for decorators
- Update Jest config to resolve @shared/types path
- All tests passing with 100% coverage

Addresses review comments on PR #5
