+++ packages/shared/package.json
@@ -7,6 +7,10 @@
     "build": "tsc",
     "dev": "tsc --watch"
   },
+  "dependencies": {
+    "class-validator": "^0.14.2",
+    "class-transformer": "^0.5.1"
+  },
   "devDependencies": {
     "typescript": "^5.3.0"
   }
